Brought type inference to 100% of the codebase (tests still missing)
Here we change the philosophy of the library radically: instead of doing runtime
validation for inputs, we leave those up to the user (yes, really! endorsing a type-checker
will need to be part of this patch), and instead we expect correct values to be passed
in as type-checked upfront.

This way, we get rid of a number of runtime validation tests that are completely avoidable.

+++ src/Formatter/BitcoinMoneyFormatter.php
@@ -61,11 +61,11 @@ final class BitcoinMoneyFormatter implements MoneyFormatter
         }
 
         if ($this->fractionDigits === 0) {
-            $formatted = substr($formatted, 0, strpos($formatted, '.'));
+            $formatted = substr($formatted, 0, (int) strpos($formatted, '.'));
         } elseif ($this->fractionDigits > $subunit) {
             $formatted .= str_pad('', $this->fractionDigits - $subunit, '0');
         } elseif ($this->fractionDigits < $subunit) {
-            $lastDigit = strpos($formatted, '.') + $this->fractionDigits + 1;
+            $lastDigit = (int) strpos($formatted, '.') + $this->fractionDigits + 1;
             $formatted = substr($formatted, 0, $lastDigit);
         }
